Register private key loading functions.
