+++ picoquic/tls_api.h
@@ -40,12 +40,6 @@
 #define PICOQUIC_LABEL_QUIC_V1_KEY_BASE "tls13 quic "
 #define PICOQUIC_LABEL_QUIC_V2_KEY_BASE "tls13 quicv2 "
 
-#define PICOQUIC_AES_128_GCM_SHA256 0x1301
-#define PICOQUIC_AES_256_GCM_SHA384 0x1302
-#define PICOQUIC_CHACHA20_POLY1305_SHA256 0x1303
-
-#define PICOQUIC_GROUP_SECP256R1 23
-
 #ifdef __cplusplus
 extern "C" {
 #endif
